+++ src/browser/entry-template.tsx
@@ -5,11 +5,9 @@ import { ExoticComponent, useEffect } from "react";
 import { createRoot } from "react-dom/client";
 import { SWRConfig } from "swr";
 
-import { getBaseFontSize, setupSentry, t } from "~/common/helpers";
+import { getBaseFontSize, t } from "~/common/helpers";
 
 import { usePreferDarkMode, useSettings } from "./hooks";
-
-setupSentry();
 
 const wrapper: EntryWrapper<ExoticComponent> = (Component) => {
   document.title = t("extensionName");